@iv-menshenin 2024-12-06
goos: linux
goarch: amd64
pkg: github.com/iv-menshenin/lyceum/collection
cpu: Intel(R) Core(TM) i7-9700F CPU @ 3.00GHz
                           │ CollectionGet_10mln │ CollectionPushPop  │  CollectionInsert   │
                           │       sec/op        │   sec/op     vs base   │    sec/op     vs base   │
*/Last-8                             1.591n ± 9%
*/First-8                            1.703n ± 4%
*/Mid-8                              1.645n ± 5%
*/Forward-8                          2.107n ± 2%
*/Backward-8                         2.364n ± 2%
*/Random-8                           3.015n ± 6%
*/Push_All-8                                       16.66n ± 8%
*/Push_Len_All-8                                   17.00n ± 7%
*/Push_Then_Pop-8                                  13.70n ± 7%
*/Push_Push_Then_Pop-8                             33.56n ± 9%
*/Push_Push_Then_Pop_Len-8                         35.47n ± 8%
*/Pop_All-8                                        7.947n ± 1%
*/Push_1K-8                                                             23.09µ ± 16%
*/Push_65K-8                                                            1.569m ± 14%
*/Push_1M-8                                                             19.81m ±  9%
*/Push_10M-8                                                            261.7m ± 30%
geomean                              2.015n        18.23n       ? ¹ ²   3.702m        ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

                           │ CollectionGet_10mln │  CollectionPushPop  │  CollectionInsert  │
                           │      allocs/op      │ allocs/op   vs base │  allocs/op   vs base   │
*/Last-8                            0.000 ± 0%
*/First-8                           0.000 ± 0%
*/Mid-8                             0.000 ± 0%
*/Forward-8                         0.000 ± 0%
*/Backward-8                        0.000 ± 0%
*/Random-8                          0.000 ± 0%
*/Push_All-8                                       0.000 ± 0%
*/Push_Len_All-8                                   0.000 ± 0%
*/Push_Then_Pop-8                                  0.000 ± 0%
*/Push_Push_Then_Pop-8                             0.000 ± 0%
*/Push_Push_Then_Pop_Len-8                         0.000 ± 0%
*/Pop_All-8                                        0.000 ± 0%
*/Push_1K-8                                                               1.000 ± 0%
*/Push_65K-8                                                              128.0 ± 0%
*/Push_1M-8                                                              1.953k ± 0%
*/Push_10M-8                                                             19.54k ± 0%
geomean                                        ¹               ? ² ¹ ³    264.4       ? ² ³
¹ summaries must be >0 to compute geomean
² benchmark set differs from baseline; geomeans may not be comparable
³ ratios must be >0 to compute geomean