@iv-menshenin 2024-12-06
```
goos: linux
goarch: amd64
pkg: github.com/iv-menshenin/lyceum/collection
cpu: Intel(R) Core(TM) i7-9700F CPU @ 3.00GHz
                           │ CollectionGet_10mln │ CollectionPushPop  │  CollectionInsert   │
                           │       sec/op        │   sec/op     vs base   │    sec/op     vs base   │
*/Last-8                             1.620n ± 3%
*/First-8                            1.694n ± 1%
*/Mid-8                              1.623n ± 4%
*/Forward-8                          2.106n ± 1%
*/Backward-8                         2.365n ± 1%
*/Random-8                           2.988n ± 1%
*/Push_All-8                                       17.33n ± 4%
*/Push_Len_All-8                                   16.85n ± 5%
*/Push_Then_Pop-8                                  13.51n ± 4%
*/Push_Push_Then_Pop-8                             33.56n ± 7%
*/Push_Push_Then_Pop_Len-8                         34.65n ± 7%
*/Pop_All-8                                        7.960n ± 0%
*/Push_1K-8                                                             23.34µ ± 14%
*/Push_65K-8                                                            1.512m ±  7%
*/Push_1M-8                                                             19.81m ±  4%
*/Push_10M-8                                                            230.1m ± 14%
geomean                              2.011n        18.21n       ? ¹ ²   3.562m        ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

                           │ CollectionGet_10mln │  CollectionPushPop  │  CollectionInsert  │
                           │      allocs/op      │ allocs/op   vs base │  allocs/op   vs base   │
*/Last-8                            0.000 ± 0%
*/First-8                           0.000 ± 0%
*/Mid-8                             0.000 ± 0%
*/Forward-8                         0.000 ± 0%
*/Backward-8                        0.000 ± 0%
*/Random-8                          0.000 ± 0%
*/Push_All-8                                       0.000 ± 0%
*/Push_Len_All-8                                   0.000 ± 0%
*/Push_Then_Pop-8                                  0.000 ± 0%
*/Push_Push_Then_Pop-8                             0.000 ± 0%
*/Push_Push_Then_Pop_Len-8                         0.000 ± 0%
*/Pop_All-8                                        0.000 ± 0%
*/Push_1K-8                                                               1.000 ± 0%
*/Push_65K-8                                                              128.0 ± 0%
*/Push_1M-8                                                              1.953k ± 0%
*/Push_10M-8                                                             19.54k ± 0%
geomean                                        ¹               ? ² ¹ ³    264.4       ? ² ³
¹ summaries must be >0 to compute geomean
² benchmark set differs from baseline; geomeans may not be comparable
³ ratios must be >0 to compute geomean
```