@iv-menshenin 2024-12-07
```
goos: linux
goarch: amd64
pkg: github.com/iv-menshenin/lyceum/collection
cpu: Intel(R) Core(TM) i7-9700F CPU @ 3.00GHz
                           │ CollectionGet_10mln │  CollectionPushPop  │  CollectionInsert   │
                           │       sec/op        │    sec/op     vs base   │    sec/op     vs base   │
*/Last-8                             1.629n ± 6%
*/First-8                            1.744n ± 1%
*/Mid-8                              1.674n ± 7%
*/Forward-8                          2.158n ± 3%
*/Backward-8                         2.384n ± 1%
*/Random-8                           3.048n ± 7%
*/Push_All-8                                       18.16n ±  8%
*/Push_Len_All-8                                   18.89n ± 12%
*/Push_Then_Pop-8                                  13.74n ±  5%
*/Push_Push_Then_Pop-8                             38.36n ±  9%
*/Push_Push_Then_Pop_Len-8                         39.75n ±  6%
*/Pop_All-8                                        7.742n ±  3%
*/Push_1K-8                                                              36.21µ ± 44%
*/Push_65K-8                                                             1.773m ± 31%
*/Push_1M-8                                                              21.33m ±  7%
*/Push_10M-8                                                             238.0m ± 50%
geomean                              2.052n        19.54n        ? ¹ ²   4.249m        ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

                           │ CollectionGet_10mln │  CollectionPushPop  │  CollectionInsert  │
                           │      allocs/op      │ allocs/op   vs base │  allocs/op   vs base   │
*/Last-8                            0.000 ± 0%
*/First-8                           0.000 ± 0%
*/Mid-8                             0.000 ± 0%
*/Forward-8                         0.000 ± 0%
*/Backward-8                        0.000 ± 0%
*/Random-8                          0.000 ± 0%
*/Push_All-8                                       0.000 ± 0%
*/Push_Len_All-8                                   0.000 ± 0%
*/Push_Then_Pop-8                                  0.000 ± 0%
*/Push_Push_Then_Pop-8                             0.000 ± 0%
*/Push_Push_Then_Pop_Len-8                         0.000 ± 0%
*/Pop_All-8                                        0.000 ± 0%
*/Push_1K-8                                                               1.000 ± 0%
*/Push_65K-8                                                              128.0 ± 0%
*/Push_1M-8                                                              1.953k ± 0%
*/Push_10M-8                                                             19.54k ± 0%
geomean                                        ¹               ? ² ¹ ³    264.4       ? ² ³
¹ summaries must be >0 to compute geomean
² benchmark set differs from baseline; geomeans may not be comparable
³ ratios must be >0 to compute geomean
```